fix(scopes): remove owner check, various cleanups

+++ controllers/common/scopes.go
@@ -26,17 +26,17 @@ func (s *scope) Process(exposedScopes map[string]naisiov1.ExposedScope) error {
 		return nil
 	}
 
-	filteredScopes, err := s.scopesExist(exposedScopes)
+	filtered, err := s.filtered(exposedScopes)
 	if err != nil {
 		return fmt.Errorf("checking if scopes exists: %w", err)
 	}
 
-	err = s.createScopes(filteredScopes.ToCreate)
+	err = s.createScopes(filtered.ToCreate)
 	if err != nil {
 		return err
 	}
 
-	err = s.updateScopes(filteredScopes.ToUpdate)
+	err = s.updateScopes(filtered.ToUpdate)
 	if err != nil {
 		return err
 	}
@@ -44,19 +44,19 @@ func (s *scope) Process(exposedScopes map[string]naisiov1.ExposedScope) error {
 	return nil
 }
 
-func (s scope) createScopes(toCreate []naisiov1.ExposedScope) error {
+func (s *scope) createScopes(toCreate []naisiov1.ExposedScope) error {
 	for _, newScope := range toCreate {
-		s.Tx.Logger.Debug(fmt.Sprintf("Subscope - %s do not exist in Digdir, creating...", newScope.Name))
+		s.Tx.Logger.Debug(fmt.Sprintf("Subscope %q does not exist in Digdir, creating...", newScope.Name))
 
 		scope, err := s.create(newScope)
 		if err != nil {
 			return err
 		}
-		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventCreatedScopeInDigDir, fmt.Sprintf("Scope created.. %s", scope.Name))
+		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventCreatedScopeInDigDir, fmt.Sprintf("Created scope %q", scope.Name))
 		metrics.IncScopesCreated(s.Tx.Instance)
 
 		// add consumers
-		_, err = s.updateConsumers(scopes.CurrentScopeInfo(*scope, newScope))
+		err = s.updateConsumers(scopes.CurrentScopeInfo(*scope, newScope))
 		if err != nil {
 			return fmt.Errorf("adding new consumers to acl: %w", err)
 		}
@@ -65,145 +65,148 @@ func (s scope) createScopes(toCreate []naisiov1.ExposedScope) error {
 	return nil
 }
 
-func (s scope) updateScopes(toUpdate []scopes.Scope) error {
+func (s *scope) updateScopes(toUpdate []scopes.Scope) error {
 	for _, scope := range toUpdate {
-		s.Tx.Logger.Debug(fmt.Sprintf("Scope - %s already exists in Digdir...", scope.ToString()))
+		s.Tx.Logger.Debug(fmt.Sprintf("updating existing scope %q...", scope.ToString()))
 
-		if scope.HasChanged() {
-			// update existing scope
-			scopeRegistration, err := s.update(scope)
-			if err != nil {
-				return err
-			}
-			s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedScopeInDigDir, fmt.Sprintf("Scope updated.. %s", scopeRegistration.Name))
-			metrics.IncScopesUpdated(s.Tx.Instance)
+		wantEnabled := scope.CurrentScope.Enabled
+		isEnabled := scope.ScopeRegistration.Active
+
+		var err error
+
+		if wantEnabled && !isEnabled {
+			err = s.activate(scope)
+		} else if wantEnabled {
+			err = s.update(scope)
+		} else {
+			err = s.deactivate(scope)
 		}
-
-		if scope.CanBeActivated() {
-			// re-activate scope
-			err := s.activate(scope)
-			if err != nil {
-				return err
-			}
-		}
-
-		if !scope.IsActive() {
-			err := s.deactivate(scope.ToString())
-			if err != nil {
-				return err
-			}
-		}
-
-		_, err := s.updateConsumers(scope)
 		if err != nil {
-			return fmt.Errorf("update consumers acl: %w", err)
+			return err
+		}
+
+		err = s.updateConsumers(scope)
+		if err != nil {
+			return fmt.Errorf("updating scope acl: %w", err)
 		}
 	}
 
 	return nil
 }
 
-func (s scope) scopesExist(exposedScopes map[string]naisiov1.ExposedScope) (*scopes.ScopeStash, error) {
-	scopeContainer, err := s.Tx.DigdirClient.GetFilteredScopes(s.Tx.Instance, s.Tx.Ctx, exposedScopes, s.Tx.Config.ClusterName)
+func (s *scope) filtered(exposedScopes map[string]naisiov1.ExposedScope) (*scopes.Operations, error) {
+	allScopes, err := s.Tx.DigdirClient.GetScopes(s.Tx.Ctx)
 	if err != nil {
-		return nil, fmt.Errorf("getting filterted scopes: %w", err)
+		return nil, fmt.Errorf("getting filtered scopes: %w", err)
 	}
-	return scopeContainer, nil
+
+	return scopes.Generate(allScopes, exposedScopes), nil
 }
 
-func (s *scope) updateConsumers(scope scopes.Scope) ([]types.ConsumerRegistration, error) {
-	s.Tx.Logger = s.Tx.Logger.WithField("scope", scope.ToString())
-	s.Tx.Logger.Debug("checking if ACL needs update...")
+func (s *scope) updateConsumers(scope scopes.Scope) error {
+	logger := s.Tx.Logger.WithField("scope", scope.ToString())
+	logger.Debug("checking if ACL needs update...")
 
 	acl, err := s.Tx.DigdirClient.GetScopeACL(s.Tx.Ctx, scope.ToString())
 	if err != nil {
-		return nil, fmt.Errorf("gettin ACL from Digdir: %w", err)
+		return fmt.Errorf("getting ACL from Digdir: %w", err)
 	}
 
 	consumerStatus, consumerList := scope.FilterConsumers(acl)
-	registrationResponse := make([]types.ConsumerRegistration, 0)
 
 	if len(consumerList) == 0 {
-		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedACLForScopeInDigDir, fmt.Sprintf("ACL was up to date for: %s", scope.ToString()))
-		return nil, nil
+		msg := fmt.Sprintf("ACL was up to date for scope %q", scope.ToString())
+		logger.Info(msg)
+		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedACLForScopeInDigDir, msg)
+		return nil
 	}
 
 	for _, consumer := range consumerList {
+		logger.Debug("processing ACLs...")
 		if consumer.ShouldBeAdded {
-			response, err := s.activateConsumer(scope.ToString(), consumer.Orgno)
+			logger.Infof("adding consumer %q to ACL...", consumer.Orgno)
+			err := s.activateConsumer(scope.ToString(), consumer.Orgno)
 			if err != nil {
-				return nil, fmt.Errorf("adding to ACL: %w", err)
+				return fmt.Errorf("adding consumer to ACL: %w", err)
 			}
+
 			consumerStatus = append(consumerStatus, consumer.Orgno)
-			registrationResponse = append(registrationResponse, *response)
 			metrics.IncScopesConsumersCreatedOrUpdated(s.Tx.Instance, consumer.State)
 		} else {
-			response, err := s.deactivateConsumer(scope.ToString(), consumer.Orgno)
+			logger.Infof("removing consumer %q from ACL...", consumer.Orgno)
+			err := s.deactivateConsumer(scope.ToString(), consumer.Orgno)
 			if err != nil {
-				return nil, fmt.Errorf("delete from ACL: %w", err)
+				return fmt.Errorf("deleting consumer from ACL: %w", err)
 			}
-			registrationResponse = append(registrationResponse, *response)
+
 			metrics.IncScopesConsumersDeleted(s.Tx.Instance)
 		}
 	}
 
-	s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedACLForScopeInDigDir, fmt.Sprintf("Scope ACL been updated.. %s", scope.ToString()))
-	return registrationResponse, nil
+	return nil
 }
 
-func (s *scope) activateConsumer(scope, consumerOrgno string) (*types.ConsumerRegistration, error) {
+func (s *scope) activateConsumer(scope, consumerOrgno string) error {
 	response, err := s.Tx.DigdirClient.AddToScopeACL(s.Tx.Ctx, scope, consumerOrgno)
 	if err != nil {
-		return nil, err
+		return err
 	}
-	msg := fmt.Sprintf("scope acl updated, added consumer: %s", consumerOrgno)
-	s.Tx.Logger.WithField("activateConsumer", response.Scope).Info(msg)
-	return response, nil
+	msg := fmt.Sprintf("Granted access to scope %q for consumer %q", scope, consumerOrgno)
+	s.Tx.Logger.WithField("scope", response.Scope).Info(msg)
+	s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedACLForScopeInDigDir, msg)
+
+	return nil
 }
 
-func (s *scope) deactivateConsumer(scope, consumerOrgno string) (*types.ConsumerRegistration, error) {
+func (s *scope) deactivateConsumer(scope, consumerOrgno string) error {
 	response, err := s.Tx.DigdirClient.DeactivateConsumer(s.Tx.Ctx, scope, consumerOrgno)
 	if err != nil {
-		return nil, err
+		return err
 	}
-	msg := fmt.Sprintf("scope acl updated, deactivated consumer: %s", consumerOrgno)
+	msg := fmt.Sprintf("Revoked access to scope %q for consumer %q", scope, consumerOrgno)
 	s.Tx.Logger.WithField("scope", response.Scope).Info(msg)
-	return response, nil
+	s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedACLForScopeInDigDir, msg)
+
+	return nil
 }
 
-func (s *scope) update(scope scopes.Scope) (*types.ScopeRegistration, error) {
+func (s *scope) update(scope scopes.Scope) error {
 	scopePayload := clients.ToScopeRegistration(s.Tx.Instance, scope.CurrentScope, s.Tx.Config)
-	s.Tx.Logger = s.Tx.Logger.WithField("scope", scope.ToString())
-	s.Tx.Logger.Debug("updating scope...")
+	s.Tx.Logger.WithField("scope", scope.ToString()).Debug("updating scope...")
 
 	registrationResponse, err := s.Tx.DigdirClient.UpdateScope(s.Tx.Ctx, scopePayload, scope.ToString())
 	if err != nil {
-		return nil, fmt.Errorf("updating scope at Digdir: %w", err)
+		return fmt.Errorf("updating scope at Digdir: %w", err)
 	}
-	return registrationResponse, err
+
+	msg := fmt.Sprintf("Updated scope %q", registrationResponse.Name)
+	s.Tx.Logger.Info(msg)
+	s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventUpdatedScopeInDigDir, msg)
+	metrics.IncScopesUpdated(s.Tx.Instance)
+
+	return nil
 }
 
 func (s *scope) create(newScope naisiov1.ExposedScope) (*types.ScopeRegistration, error) {
-	scopeRegistrationPayload := clients.ToScopeRegistration(s.Tx.Instance, newScope, s.Tx.Config)
+	payload := clients.ToScopeRegistration(s.Tx.Instance, newScope, s.Tx.Config)
 	s.Tx.Logger.Debug("scope does not exist in Digdir, registering...")
 
-	registrationResponse, err := s.Tx.DigdirClient.RegisterScope(s.Tx.Ctx, scopeRegistrationPayload)
+	response, err := s.Tx.DigdirClient.RegisterScope(s.Tx.Ctx, payload)
 	if err != nil {
 		return nil, fmt.Errorf("registering client to Digdir: %w", err)
 	}
 
-	s.Tx.Logger = s.Tx.Logger.WithField("scope", registrationResponse.Name)
-	s.Tx.Logger.Info("scope registered")
-	return registrationResponse, nil
+	s.Tx.Logger.WithField("scope", response.Name).Info("scope registered")
+	return response, nil
 }
 
-func (s *scope) deactivate(scope string) error {
-	scopeRegistration, err := s.Tx.DigdirClient.DeleteScope(s.Tx.Ctx, scope)
+func (s *scope) deactivate(scope scopes.Scope) error {
+	registration, err := s.Tx.DigdirClient.DeleteScope(s.Tx.Ctx, scope.ToString())
 	if err != nil {
 		return fmt.Errorf("deleting scope: %w", err)
 	}
 
-	msg := fmt.Sprintf("Scope '%s' deactivated, consumers no longer have access", scopeRegistration.Name)
+	msg := fmt.Sprintf("Deactivated scope %q; consumers no longer have access", registration.Name)
 	s.Tx.Logger.Warning(msg)
 	s.Rec.reportEvent(s.Tx, corev1.EventTypeWarning, EventDeactivatedScopeInDigDir, msg)
 	metrics.IncScopesDeleted(s.Tx.Instance)
@@ -212,13 +215,13 @@ func (s *scope) deactivate(scope string) error {
 }
 
 func (s *scope) activate(scope scopes.Scope) error {
-	scopeActivationPayload := clients.ToScopeRegistration(s.Tx.Instance, scope.CurrentScope, s.Tx.Config)
-	scopeRegistration, err := s.Tx.DigdirClient.ActivateScope(s.Tx.Ctx, scopeActivationPayload, scope.ToString())
+	payload := clients.ToScopeRegistration(s.Tx.Instance, scope.CurrentScope, s.Tx.Config)
+	registration, err := s.Tx.DigdirClient.UpdateScope(s.Tx.Ctx, payload, scope.ToString())
 	if err != nil {
 		return fmt.Errorf("activating scope: %w", err)
 	}
 
-	msg := fmt.Sprintf("Scope '%s' activated", scopeRegistration.Name)
+	msg := fmt.Sprintf("Activated scope %q", registration.Name)
 	s.Tx.Logger.Info(msg)
 	s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventActivatedScopeInDigDir, msg)
 	metrics.IncScopesReactivated(s.Tx.Instance)
@@ -227,7 +230,7 @@ func (s *scope) activate(scope scopes.Scope) error {
 }
 
 func (s *scope) Finalize(exposedScopes map[string]naisiov1.ExposedScope) error {
-	filteredScopes, err := s.Tx.DigdirClient.GetFilteredScopes(s.Tx.Instance, s.Tx.Ctx, exposedScopes, s.Tx.Config.ClusterName)
+	filteredScopes, err := s.filtered(exposedScopes)
 	if err != nil {
 		return err
 	}
@@ -237,12 +240,17 @@ func (s *scope) Finalize(exposedScopes map[string]naisiov1.ExposedScope) error {
 	}
 
 	for _, scope := range filteredScopes.ToUpdate {
-		s.Tx.Logger.Infof("delete annotation set, deleting scope: %s from Maskinporten... ", scope.ToString())
+		if scope.CurrentScope.Enabled {
+			s.Tx.Logger.Infof("scope %q is still set to enabled, skipping deletion... ", scope.ToString())
+			continue
+		}
+
+		s.Tx.Logger.Infof("finalizer triggered, deleting scope %q from Maskinporten... ", scope.ToString())
 		if _, err := s.Tx.DigdirClient.DeleteScope(s.Tx.Ctx, scope.ToString()); err != nil {
 			return fmt.Errorf("deactivate scope in Maskinporten: %w", err)
 		}
 		metrics.IncScopesDeleted(s.Tx.Instance)
-		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventDeactivatedScopeInDigDir, "Scope deactivated in Digdir")
+		s.Rec.reportEvent(s.Tx, corev1.EventTypeNormal, EventDeactivatedScopeInDigDir, fmt.Sprintf("Deactivated scope %q", scope.ToString()))
 	}
 
 	return nil
